backend/requirements.txt
fastapi==0.128.3
uvicorn==0.40.0
requests==2.32.3
python-multipart==0.0.22
faiss-cpu>=1.7.4
numpy>=1.24.0
PyMuPDF>=1.24.0
